# api/docs/swagger.yaml
basePath: /api
definitions:
...
  dto.AccountDTO:
    properties:
      category:
        type: string
      id:
        type: string
      name:
        type: string
      role:
        type: string
    type: object
  dto.AuthResponse:
    properties:
      accessToken:
        type: string
      accounts:
        items:
          $ref: '#/definitions/dto.AccountDTO'
        type: array
      refreshToken:
        type: string
      user:
        $ref: '#/definitions/dto.UserResponse'
    type: object
...
  dto.ChangePasswordDTO:
    properties:
      newPassword:
        minLength: 6
        type: string
      oldPassword:
        type: string
    required:
    - newPassword
    - oldPassword
    type: object
...
  dto.ForgotPasswordDTO:
    properties:
      identifier:
        type: string
    required:
    - identifier
    type: object
...
  dto.InviteUserDTO:
    properties:
      email:
        type: string
      phone:
        maxLength: 15
        minLength: 10
        type: string
      role:
        enum:
        - owner
        - manager
        - staff
        - admin
        type: string
    required:
    - role
    type: object
  dto.JoinTenantDTO:
    properties:
      name:
        maxLength: 100
        minLength: 2
        type: string
      password:
        minLength: 6
        type: string
      token:
        type: string
    required:
    - name
    - password
    - token
    type: object
...
  dto.LoginDTO:
    properties:
      password:
        type: string
      phone:
        type: string
    required:
    - password
    - phone
    type: object
  dto.LogoutDTO:
    properties:
      refreshToken:
        type: string
    type: object
  dto.Pagination:
    properties:
      currentPage:
        type: integer
      limit:
        type: integer
      totalItems:
        type: integer
      totalPages:
        type: integer
    type: object
...
  dto.RefreshTokenDTO:
    properties:
      refreshToken:
        type: string
    required:
    - refreshToken
    type: object
  dto.RegisterIndividualDTO:
    properties:
      email:
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
      password:
        maxLength: 50
        minLength: 6
        type: string
      phone:
        maxLength: 15
        minLength: 10
        type: string
    required:
    - name
    - password
    - phone
    type: object
  dto.RegisterTenantDTO:
    properties:
      email:
        type: string
      ownerName:
        maxLength: 100
        minLength: 2
        type: string
      password:
        maxLength: 50
        minLength: 6
        type: string
      phone:
        maxLength: 15
        minLength: 10
        type: string
      tenantName:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - ownerName
    - password
    - phone
    - tenantName
    type: object
  dto.ResetPasswordDTO:
    properties:
      identifier:
        type: string
      newPassword:
        maxLength: 50
        minLength: 6
        type: string
      otp:
        type: string
    required:
    - identifier
    - newPassword
    - otp
    type: object
...
  dto.UserListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.UserResponse'
        type: array
      pagination:
        $ref: '#/definitions/dto.Pagination'
    type: object
  dto.UserResponse:
    properties:
      createdAt:
        type: string
      email:
        type: string
      id:
        type: string
      isActive:
        type: boolean
      name:
        type: string
      phone:
        type: string
      role:
        type: string
      tenantId:
        type: string
    type: object
  dto.VerifyOTPDTO:
    properties:
      otp:
        type: string
      phone:
        type: string
    required:
    - otp
    - phone
    type: object
...
  handler.CreatePlanRequest:
    properties:
      code:
        type: string
      description:
        type: string
      features:
        additionalProperties:
          type: boolean
        type: object
      interval:
        enum:
        - MONTHLY
        - YEARLY
        type: string
      limits:
        additionalProperties:
          type: integer
        type: object
      name:
        type: string
      price:
        minimum: 0
        type: number
    required:
    - code
    - name
    type: object
...
  handler.SubscribeRequest:
    properties:
      password:
        type: string
      planId:
        type: string
    required:
    - password
    - planId
    type: object
...
host: localhost:4000
info:
  contact: {}
  description: High performance Go API for AceExtension
  title: AceExtension API
  version: "1.0"
paths:
...
  /api/v1/plans:
    get:
      description: List all available subscription plans
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List plans
      tags:
      - plans
    post:
      consumes:
      - application/json
      description: Create a new subscription plan (Admin)
      parameters:
      - description: Plan Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreatePlanRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a plan
      tags:
      - plans
...
  /api/v1/subscriptions/current:
    get:
      description: Get usage and subscription details for the tenant
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get current subscription
      tags:
      - subscriptions
  /api/v1/subscriptions/subscribe:
    post:
      consumes:
      - application/json
      description: Upgrade or change subscription plan
      parameters:
      - description: Subscribe Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.SubscribeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Subscribe to plan
      tags:
      - subscriptions
...
  /auth/change-password:
    post:
      consumes:
      - application/json
      description: Change authenticated user password
      parameters:
      - description: Change Password Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ChangePasswordDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Change Password
      tags:
      - auth
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Request OTP for password reset
      parameters:
      - description: Forgot Password Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ForgotPasswordDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Forgot Password
      tags:
      - auth
  /auth/impersonate/{tenantId}:
    post:
      consumes:
      - application/json
      description: Login as tenant owner (Super Admin only)
      parameters:
      - description: Tenant ID
        in: path
        name: tenantId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AuthResponse'
      summary: Impersonate Tenant
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login with phone or email and password
      parameters:
      - description: Login Credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AuthResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: User Login
      tags:
      - auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Invalidate the session
      parameters:
      - description: Logout Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LogoutDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: User Logout
      tags:
      - auth
  /auth/me:
    get:
      consumes:
      - application/json
      description: Get profile of the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get Current User
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Get a new access token using refresh token
      parameters:
      - description: Refresh Token Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RefreshTokenDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AuthResponse'
      summary: Refresh Access Token
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a new tenant and its administrative user
      parameters:
      - description: Tenant Registration Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterTenantDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new tenant
      tags:
      - auth
  /auth/register-individual:
    post:
      consumes:
      - application/json
      description: Create a new personal account and its owner
      parameters:
      - description: Individual Registration Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterIndividualDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new individual user
      tags:
      - auth
  /auth/reset-password:
    post:
      consumes:
      - application/json
      description: Reset password with OTP
      parameters:
      - description: Reset Password Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ResetPasswordDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Reset Password
      tags:
      - auth
  /auth/verify-otp:
    post:
      consumes:
      - application/json
      description: Verify the 6-digit OTP sent during registration
      parameters:
      - description: OTP Verification Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.VerifyOTPDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AuthResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Verify registration OTP
      tags:
      - auth
  /users:
    get:
      consumes:
      - application/json
      description: Fetch users with advanced filtering, sorting and pagination
      parameters:
      - description: Search term
        in: query
        name: search
        type: string
      - description: Field to sort by
        in: query
        name: sortBy
        type: string
      - description: Sort order (asc/desc)
        in: query
        name: sortOrder
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      - description: JSON encoded filters
        in: query
        name: filters
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserListResponse'
      security:
      - BearerAuth: []
      summary: List tenant users
      tags:
      - users
  /users/invite:
    post:
      consumes:
      - application/json
      description: Send an invitation to join the tenant
      parameters:
      - description: Invitation Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.InviteUserDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Invite a new user
      tags:
      - users
  /users/join:
    post:
      consumes:
      - application/json
      description: Accept invitation and create user account
      parameters:
      - description: Join Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.JoinTenantDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Join tenant via invitation
      tags:
      - users
swagger: "2.0"